Add API for extended audit-logs sidecar settings
